LU-10973 lutf: Fix crash and other updates
Fix crash in wait_for_agents. Was mis-using
cYAML_get_next_seq_item().

Update the lustre_lnet_config_ni() with a newly added parameter
for conns_per_peer. Later on tests can be added to explicitly
test setting the conns_per_peer from the C API.

Remove auth_timeout from the paramiko file to be backwards
compatible with older versions of the paramiko python API.

Only delete the progress file if this node is the LUTF master
node. This is to avoid other nodes trampling over each other
if they are using the same directory to dump temporary files.

Test-parameters: trivial

Reviewed-on: https://review.whamcloud.com/44726
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/tests/lutf/python/infra/lutf.py
@@ -770,7 +770,7 @@ class Myself:
 			self.__lustre_base_path = ''
 		self.alias_list = self.provision_intfs(config_ifs_num)
 		# delete any older test_progress files
-		if os.path.isfile(self.get_test_progress_path()):
+		if os.path.isfile(self.get_test_progress_path()) and self.__lutf_type == EN_LUTF_MASTER:
 			os.remove(self.get_test_progress_path())
 
 	def import_env_vars(self, fpath):
